move video gallery jsx to a separate module
